Move templates back into compiler
