+++ packages/compiler/.mocharc.yaml
@@ -1,6 +1,6 @@
 timeout: 5000
 require: source-map-support/register
-spec: "dist/test/**/*.{test,e2e}.js"
+spec: "dist/test/**/*.test.js"
 ignore: "dist/test/manual/**/*.js"
 
 # Config for https://www.npmjs.com/package/mocha-multi-reporters
