+++ packages/compiler/templates/emitter-ts/prettierrc.yaml
