--- packages/init-templates/.eslintignore
@@ -1,3 +1,0 @@
-templates/
-temp/
-dist/
