--- packages/init-templates/mocha.reporter.config.json
@@ -1,4 +1,0 @@
-{
-  "reporterEnabled": "spec, mocha-junit-reporter",
-  "maxDiffSize": 16384
-}
