--- packages/init-templates/tsconfig.build.json
@@ -1,9 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "outDir": "dist",
-    "rootDir": "src",
-    "tsBuildInfoFile": "temp/tsconfig.tsbuildinfo"
-  },
-  "include": ["src/**/*.ts"]
-}
